joyca/requirements.txt
python-decouple==3.8
Django==5.0.3
mysqlclient==2.2.4
requests==2.31.0
ntscraper==0.3.13
tqdm
django-import-export==3.3.8
django-multi-domains